[AI] Extract work with designs from production to separate module
- Make design cache calculation lazy, no need to call it during turn preparation explicitly (remove temporal coupling)
- Separate function to two parts: outer is to normalize arguments and shape return type and inner to do data aggregation staff,
- Remove unused get_design_cost
- Use a new approach to import from submodule: `from turn_state.design import ...`, this allows to handle circular imports,
- Annotate arguments and return types, update usages
