+++ default/python/AI/ProductionAI.py
@@ -2,7 +2,7 @@ import math
 import random
 from logging import debug, error, info, warning
 from operator import itemgetter
-from typing import Iterable, List, Tuple
+from typing import List
 
 import freeOrionAIInterface as fo  # pylint: disable=import-error
 
@@ -13,176 +13,22 @@ import FleetUtilsAI
 import MilitaryAI
 import PlanetUtilsAI
 import PriorityAI
-import ShipDesignAI
 from AIDependencies import INVALID_ID
 from EnumsAI import (EmpireProductionTypes, FocusType, MissionType, PriorityType, ShipRoleType,
                      get_priority_production_types, )
 from aistate_interface import get_aistate
 from character.character_module import Aggression
 from common.print_utils import Sequence, Table, Text
-from freeorion_tools import AITimer, chat_human, ppstring, tech_is_complete
+from freeorion_tools import chat_human, ppstring, tech_is_complete
 from turn_state import (best_pilot_rating, get_all_empire_planets, get_empire_drydocks, get_empire_outposts,
                         get_empire_planets_by_species,
                         get_inhabited_planets, medium_pilot_rating, get_owned_planets_in_system, get_owned_planets, get_colonized_planets,
                         get_empire_planets_with_species,
                         population_with_industry_focus, )
+from turn_state.design import get_best_ship_info, get_best_ship_ratings
 
-_best_military_design_rating_cache = {}  # indexed by turn, values are rating of the military design of the turn
-_design_cost_cache = {0: {(-1, -1): 0}}  # outer dict indexed by cur_turn (currently only one turn kept); inner dict indexed by (design_id, pid)
-
-_design_cache = {}  # dict of tuples (rating,pid,designID,cost) sorted by rating and indexed by priority type
 
 _CHAT_DEBUG = False
-
-
-def find_best_designs_this_turn():
-    """Calculate the best designs for each ship class available at this turn."""
-    design_timer = AITimer('ShipDesigner')
-    design_timer.start('Updating cache for new turn')
-    ShipDesignAI.Cache.update_for_new_turn()
-    _design_cache.clear()
-
-    # TODO Dont use PriorityType but introduce more reasonable Enum
-    designers = [
-        ('Orbital Invasion', PriorityType.PRODUCTION_ORBITAL_INVASION, ShipDesignAI.OrbitalTroopShipDesigner),
-        ('Invasion', PriorityType.PRODUCTION_INVASION, ShipDesignAI.StandardTroopShipDesigner),
-        ('Orbital Colonization', PriorityType.PRODUCTION_ORBITAL_COLONISATION, ShipDesignAI.OrbitalColonisationShipDesigner),
-        ('Colonization', PriorityType.PRODUCTION_COLONISATION, ShipDesignAI.StandardColonisationShipDesigner),
-        ('Orbital Outposter', PriorityType.PRODUCTION_ORBITAL_OUTPOST, ShipDesignAI.OrbitalOutpostShipDesigner),
-        ('Outposter', PriorityType.PRODUCTION_OUTPOST, ShipDesignAI.StandardOutpostShipDesigner),
-        ('Orbital Defense', PriorityType.PRODUCTION_ORBITAL_DEFENSE, ShipDesignAI.OrbitalDefenseShipDesigner),
-        ('Scouts', PriorityType.PRODUCTION_EXPLORATION, ShipDesignAI.ScoutShipDesigner),
-    ]
-
-    for timer_name, priority_type, designer in designers:
-        design_timer.start(timer_name)
-        _design_cache[priority_type] = designer().optimize_design()
-    best_military_stats = ShipDesignAI.WarShipDesigner().optimize_design()
-    best_carrier_stats = ShipDesignAI.CarrierShipDesigner().optimize_design()
-    best_stats = best_military_stats + best_carrier_stats if random.random() < .8 else best_military_stats
-    best_stats.sort(reverse=True)
-    _design_cache[PriorityType.PRODUCTION_MILITARY] = best_stats
-    design_timer.start('Krill Spawner')
-    ShipDesignAI.KrillSpawnerShipDesigner().optimize_design()  # just designing it, building+mission not supported yet
-    if fo.currentTurn() % 10 == 0:
-        design_timer.start('Printing')
-        ShipDesignAI.Cache.print_best_designs()
-    design_timer.stop_print_and_clear()
-
-
-def get_design_cost(design: "fo.shipDesign", pid: int) -> float:  # TODO: Use new framework
-    """
-    Find and return the design_cost of the specified design on the specified planet.
-    """
-    cur_turn = fo.currentTurn()
-    if cur_turn in _design_cost_cache:
-        cost_cache = _design_cost_cache[cur_turn]
-    else:
-        _design_cost_cache.clear()
-        cost_cache = {}
-        _design_cost_cache[cur_turn] = cost_cache
-    loc_invariant = design.costTimeLocationInvariant
-    if loc_invariant:
-        loc = INVALID_ID
-    else:
-        loc = pid
-    return cost_cache.setdefault((design.id, loc), design.productionCost(fo.empireID(), pid))
-
-
-def cur_best_military_design_rating() -> float:
-    """
-    Find and return the default combat rating of our best military design.
-    """
-    current_turn = fo.currentTurn()
-    if current_turn in _best_military_design_rating_cache:
-        return _best_military_design_rating_cache[current_turn]
-    priority = PriorityType.PRODUCTION_MILITARY
-    if _design_cache.get(priority, None) and _design_cache[priority][0]:
-        # the rating provided by the ShipDesigner does not
-        # reflect the rating used in threat considerations
-        # but takes additional factors (such as cost) into
-        # account. Therefore, we want to calculate the actual
-        # rating of the design as provided by CombatRatingsAI.
-        _, _, _, _, stats = _design_cache[priority][0]
-        # TODO: Should this consider enemy stats?
-        rating = stats.convert_to_combat_stats().get_rating()
-        _best_military_design_rating_cache[current_turn] = rating
-        return max(rating, 0.001)
-    return 0.001
-
-
-def get_best_ship_info(priority, loc=None):
-    """ Returns 3 item tuple: designID, design, buildLocList."""
-    if loc is None:
-        planet_ids = get_inhabited_planets()
-    elif isinstance(loc, list):
-        planet_ids = set(loc).intersection(get_inhabited_planets())
-    elif isinstance(loc, int) and loc in get_inhabited_planets():
-        planet_ids = [loc]
-    else:  # problem
-        return None, None, None
-    if priority in _design_cache:
-        best_designs = _design_cache[priority]
-        if not best_designs:
-            return None, None, None
-
-        # best_designs are already sorted by rating high to low, so the top rating is the first encountered within
-        # our planet search list
-        for design_stats in best_designs:
-            top_rating, pid, top_id, cost, stats = design_stats
-            if pid in planet_ids:
-                break
-        else:
-            return None, None, None  # apparently can't build for this priority within the desired planet group
-        valid_locs = [pid_ for rating, pid_, design_id, _, _ in best_designs if
-                      rating == top_rating and design_id == top_id and pid_ in planet_ids]
-        return top_id, fo.getShipDesign(top_id), valid_locs
-    else:
-        return None, None, None  # must be missing a Shipyard or other orbital (or missing tech)
-
-
-def get_best_ship_ratings(planet_ids: Iterable[int]) -> List[Tuple[float, int, int, "fo.shipDesign"]]:
-    """
-    Returns list of [partition, pid, designID, design] sublists, currently only for military ships.
-
-    Since we haven't yet implemented a way to target military ship construction at/near particular locations
-    where they are most in need, and also because our rating system is presumably useful-but-not-perfect, we want to
-    distribute the construction across the Resource Group and across similarly rated designs, preferentially choosing
-    the best rated design/loc combo, but if there are multiple design/loc combos with the same or similar ratings then
-    we want some chance of choosing  those alternate designs/locations.
-
-    The approach to this taken below is to treat the ratings akin to an energy to be used in a statistical mechanics
-    type partition function. 'tally' will compute the normalization constant.
-    So first go through and calculate the tally as well as convert each individual contribution to
-    the running total up to that point, to facilitate later sampling.  Then those running totals are
-    renormalized by the final tally, so that a later random number selector in the range [0,1) can be
-    used to select the chosen design/loc.
-
-    :param planet_ids: list of planets ids.
-    """
-    priority = PriorityType.PRODUCTION_MILITARY
-    planet_ids = set(planet_ids).intersection(ColonisationAI.empire_shipyards)
-
-    if priority in _design_cache:
-        build_choices = _design_cache[priority]
-        loc_choices = [[rating, pid, design_id, fo.getShipDesign(design_id)]
-                       for (rating, pid, design_id, cost, stats) in build_choices if pid in planet_ids]
-        if not loc_choices:
-            return []
-        best_rating = loc_choices[0][0]
-        tally = 0
-        ret_val = []
-        for rating, pid, design_id, design in loc_choices:
-            if rating < 0.7 * best_rating:
-                break
-            p = math.exp(10 * (rating/best_rating - 1))
-            tally += p
-            ret_val.append([tally, pid, design_id, design])
-        for item in ret_val:
-            item[0] /= tally
-        return ret_val
-    else:
-        return []
 
 
 # TODO Move Building names to AIDependencies to avoid typos and for IDE-Support
@@ -1281,7 +1127,7 @@ def generate_production_orders():
         debug("%.2f PP remaining in system group: %s", avail_pp, PlanetUtilsAI.sys_name_ids(set(PlanetUtilsAI.get_systems(planet_set))))
         debug("\t owned planets in this group are:")
         debug("\t %s", PlanetUtilsAI.planet_string(planet_set))
-        best_design_id, best_design, build_choices = get_best_ship_info(PriorityType.PRODUCTION_COLONISATION, list(planet_set))
+        best_design_id, best_design, build_choices = get_best_ship_info(PriorityType.PRODUCTION_COLONISATION, planet_set)
         species_map = {}
         for loc in (build_choices or []):
             this_spec = universe.getPlanet(loc).speciesName
@@ -1304,7 +1150,7 @@ def generate_production_orders():
                 # score = ColonisationAI.pilotRatings.get(pid, 0)
                 # if bestScore < ColonisationAI.curMidPilotRating:
             else:
-                best_design_id, best_design, build_choices = get_best_ship_info(priority, list(planet_set))
+                best_design_id, best_design, build_choices = get_best_ship_info(priority, planet_set)
             if best_design is None:
                 del local_priorities[priority]  # must be missing a shipyard -- TODO build a shipyard if necessary
                 continue
